requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
redis>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
daphne>=4.0,<5.0
Pillow>=10.0,<11.0
django-filter>=23.0,<24.0
bleach>=6.0,<7.0
whitenoise>=6.0,<7.0
django-redis>=6.0,<7.0
gunicorn>=21.0,<22.0
uvicorn[standard]>=0.27,<1.0
